Fix the build
The dev server was running (with warnings) but the build script failed
to complete with the same render-spy JSX errors that I was seeing while
running the dev server. I haven't exactly found clear guidelines on
this, but it seems like I need to independently add the @react/types for
typescript to work -- but maybe that's temporary, or maybe if I removed
render-spy I wouldn't have to?