Implement basic upload
When the user snaps a photo, it gets uploaded to firebase. There's no
way to see it again and the photos is always uploaded to the same
location -- plus it's accessible to anybody who's authenticated and the
upload doesn't overwrite automatically, so it errors if the old file
hasn't been cleared out first, but.... it's a start.