Quick fix for navbar on mobile
The menu doesn't fit on a typical mobile form factor as is, so I'm
hiding the app name in the header for small screens. Need to come up
with a better fix, but I wanted to move on to trying to package the
current state up with cordova before doing much more with the app logic
itself.