Modify preact build to work with cordova webviews
Cordova just wraps a webapp up into an Android or iOS webview and then
packages that as an app. The main challenge here is that the webview
sources all of the assets as file URLs which means a few things:

1. ESM modules are no longer an option because they elevate a number of
security restrictions that break everything
2. Consistently referring to files relative to the location of the root
"www" folder becomes tricky -- can't simply rely on "/" as the base url
because for a file url that would be file:/// which is not necessarily
the folder that hosts the root of the webapp (the entry level index.html
file) -- the accepted hack here is to use "" as the publicPath setting
in webpack, basically using relative paths to css and js resources from
the main index.html file
3. Routes are tricky to manage since we can't any longer rely on the
base "/" route -- using the HashRouter solves this -- routes come after
the # in the url and they're essentially treated like anchor links.
4. Cordova expects everything to be in the www folder, not the build
folder

After creating a separate build that takes advantage of all of the
above, you can actually open the www/index.html file coming out of the
cordova build (`npm run build:cordova`) and use it without running an
http server.