Install cordova
I've installed the main npm cordova dependencies and a few plugins that
will probably be needed for firebase and auth, but for now I just wanted
to commit the minimal amount of work necessary to get the app up and
running. So far I am NOT including anything from the plugin or platform
directories as those can be fully recreated using:

    npx cordova prepare

Once the platforms and plugins have been programmatically generated the
app itself can be built with:

    npx cordova build

And the resulting debug apk can be deployed to an emulator. It works but
fails as expected when I try to log in with an error mentioning that the
universal link plugin workaround hasn't been installed.