Add last cordova plugin necessary for android firebase auth
It wasn't clear from the documentation if we actually needed this one
(it said "to support older devices") but apparently we did. I've added
it and started updating the HTML Content Security Protocol header lines.
This is just a start, I'm already getting more errors after adding this
one, but I don't want to allow wildcard access -- I'd rather try to
figure out exactly who we need to let through one which protocols.