Configure jest and enzyme for basic tests
The header test that came with the typescript template didn't actually
run. I don't know that this is the best way to do tests, but it's a
start for now. I'm about to start implementing some features and I'd
actually like to use it to help with my development, so I'll adjust it
as needed as I go.